ms-operation-game/requirements.txt
fastapi
uvicorn
sqlmodel
python-dotenv
psycopg2-binary
PyJWT
python-multipart